Pyhou/pyproject.toml
[project]
name = "pyhou"
version = "0.1.0"
description = "Touhou in python"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pygame-ce>=2.5.5",
]
